requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
python-dotenv==1.0.1

pandas==2.2.1
pyarrow==15.0.0
matplotlib==3.8.3
seaborn==0.13.2

streamlit==1.32.2
plotly==5.20.0

pytest==8.1.1
pytest-cov==5.0.0
